backend/requirements.txt
# Core web framework
fastapi

# ASGI server for running FastAPI
uvicorn[standard]

# Pydantic (often comes with FastAPI, but explicit is good)
pydantic
pydantic-settings # For loading config from env vars

# Google Cloud Libraries
google-cloud-firestore
google-cloud-storage

# Firebase Authentication
firebase-admin>=6.4.0

# For reading .env files during local development
python-dotenv

# For generating unique IDs
uuid

# Used for timezone name extraction
pytz

# For handling multipart/form-data (file uploads)
python-multipart

# For JWT authentication
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
